requirements.txt
anyio==4.10.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.1
python-telegram-bot==20.3
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
